Remove duplication between base58 and base32 wiring
