+++ src/Functions/base32Decode.cpp
@@ -11,8 +11,9 @@ struct NameBase32Decode
     static constexpr auto name = "base32Decode";
 };
 
-using Base32DecodeImpl = Base32Decode<NameBase32Decode, Base32DecodeErrorHandling::ThrowException>;
-using FunctionBase32Decode = FunctionBase32Conversion<Base32DecodeImpl>;
+
+using Base32DecodeImpl = BaseXXDecode<Base32Traits, NameBase32Decode, BaseXXDecodeErrorHandling::ThrowException>;
+using FunctionBase32Decode = FunctionBaseXXConversion<Base32DecodeImpl>;
 
 }
 
@@ -20,5 +21,4 @@ REGISTER_FUNCTION(Base32Decode)
 {
     factory.registerFunction<FunctionBase32Decode>();
 }
-
 }
