+++ src/Functions/base58Decode.cpp
@@ -11,8 +11,9 @@ struct NameBase58Decode
     static constexpr auto name = "base58Decode";
 };
 
-using Base58DecodeImpl = Base58Decode<NameBase58Decode, Base58DecodeErrorHandling::ThrowException>;
-using FunctionBase58Decode = FunctionBase58Conversion<Base58DecodeImpl>;
+
+using Base58DecodeImpl = BaseXXDecode<Base58Traits, NameBase58Decode, BaseXXDecodeErrorHandling::ThrowException>;
+using FunctionBase58Decode = FunctionBaseXXConversion<Base58DecodeImpl>;
 
 }
